ref: remove repositories from controllers and add dto

+++ services/backend/product-service/src/main/java/com/veteroch4k/product/models/ProductManufacturingInfo.java
@@ -1,9 +1,9 @@
 package com.veteroch4k.product.models;
 
 public record ProductManufacturingInfo(
-    Integer productId,
+    Long productId,
     String description,
-    Integer drawingId,
-    Integer factoryId,
-    Integer operationId
+    Long drawingId,
+    Long factoryId,
+    Long operationId
 ) {}
